requirements.txt
prettytable==3.12.0
requests==2.32.3
python-dotenv==1.0.1